Extract lastUpdatedTimeLayergroup middleware
