+++ lib/cartodb/controllers/map/anonymous.js
@@ -15,6 +15,7 @@ const cacheControlHeader = require('../../middleware/cache-control-header');
 const cacheChannelHeader = require('../../middleware/cache-channel-header');
 const surrogateKeyHeader = require('../../middleware/surrogate-key-header');
 const lastModifiedHeader = require('../../middleware/last-modified-header');
+const lastUpdatedTimeLayergroup = require('./middlewares/last-updated-time-layergroup');
 const sendResponse = require('../../middleware/send-response');
 const CreateLayergroupMapConfigProvider = require('../../models/mapconfig/provider/create-layergroup-provider');
 const LayergroupMetadata = require('../../utils/layergroup-metadata');
@@ -99,7 +100,7 @@ AnonymousMapController.prototype.composeCreateMapMiddleware = function () {
         cacheChannelHeader(),
         surrogateKeyHeader({ surrogateKeysCache: this.surrogateKeysCache }),
         lastModifiedHeader({ now: true }),
-        setLastUpdatedTimeToLayergroup(),
+        lastUpdatedTimeLayergroup(),
         setLayerStats(this.pgConnection, this.statsBackend),
         setLayergroupIdHeader(this.templateMaps ,useTemplateHash),
         setDataviewsAndWidgetsUrlsToLayergroupMetadata(this.layergroupMetadata),
@@ -214,46 +215,6 @@ function createLayergroup (mapBackend, userLimitsApi, pgConnection, affectedTabl
             next();
         });
     };
-}
-
-function setLastUpdatedTimeToLayergroup () {
-    return function setLastUpdatedTimeToLayergroupMiddleware (req, res, next) {
-        const { mapConfigProvider, analysesResults } = res.locals;
-        const layergroup = res.body;
-
-        mapConfigProvider.getAffectedTables((err, affectedTables) => {
-            if (err) {
-                return next(err);
-            }
-
-            if (!affectedTables) {
-                return next();
-            }
-
-            var lastUpdateTime = affectedTables.getLastUpdatedAt();
-
-            lastUpdateTime = getLastUpdatedTime(analysesResults, lastUpdateTime) || lastUpdateTime;
-
-            // last update for layergroup cache buster
-            layergroup.layergroupid = layergroup.layergroupid + ':' + lastUpdateTime;
-            layergroup.last_updated = new Date(lastUpdateTime).toISOString();
-
-            next();
-        });
-    };
-}
-
-function getLastUpdatedTime(analysesResults, lastUpdateTime) {
-    if (!Array.isArray(analysesResults)) {
-        return lastUpdateTime;
-    }
-    return analysesResults.reduce(function(lastUpdateTime, analysis) {
-        return analysis.getNodes().reduce(function(lastNodeUpdatedAtTime, node) {
-            var nodeUpdatedAtDate = node.getUpdatedAt();
-            var nodeUpdatedTimeAt = (nodeUpdatedAtDate && nodeUpdatedAtDate.getTime()) || 0;
-            return nodeUpdatedTimeAt > lastNodeUpdatedAtTime ? nodeUpdatedTimeAt : lastNodeUpdatedAtTime;
-        }, lastUpdateTime);
-    }, lastUpdateTime);
 }
 
 function setLayerStats (pgConnection, statsBackend) {
